+++ tests/unit/test_architecture_adapter.py
@@ -3,58 +3,15 @@
 import pytest
 import torch.nn as nn
 
-from tests.mocks.models import MockGemma3Model
-from transformer_lens.model_bridge.architecture_adapter import ArchitectureAdapter
-from transformer_lens.model_bridge.generalized_components import (
-    AttentionBridge,
-    BlockBridge,
-    LayerNormBridge,
-    MLPBridge,
+from tests.mocks.architecture_adapter import (
+    MockArchitectureAdapter,
+    mock_adapter,
+    mock_model,
 )
+from tests.mocks.models import MockGemma3Model
 from transformer_lens.model_bridge.supported_architectures.gemma3 import (
     Gemma3ArchitectureAdapter,
 )
-
-
-class MockArchitectureAdapter(ArchitectureAdapter):
-    """Mock architecture adapter for testing."""
-
-    def __init__(self, cfg=None):
-        super().__init__(cfg)
-        self.component_mapping = {
-            "ln_final": ("ln_final", LayerNormBridge),
-            "blocks": (
-                "blocks",
-                BlockBridge,
-                {
-                    "ln1": ("ln1", LayerNormBridge),
-                    "ln2": ("ln2", LayerNormBridge),
-                    "attn": ("attn", AttentionBridge),
-                    "mlp": ("mlp", MLPBridge),
-                },
-            ),
-        }
-
-
-@pytest.fixture
-def mock_adapter() -> MockArchitectureAdapter:
-    """Create a mock adapter."""
-    return MockArchitectureAdapter()
-
-
-@pytest.fixture
-def mock_model() -> nn.Module:
-    """Create a mock model for testing."""
-    model = nn.Module()
-    model.ln_final = nn.LayerNorm(10)
-    model.blocks = nn.ModuleList()
-    block = nn.Module()
-    block.ln1 = nn.LayerNorm(10)
-    block.ln2 = nn.LayerNorm(10)
-    block.attn = nn.Module()
-    block.mlp = nn.Module()
-    model.blocks.append(block)
-    return model
 
 
 def test_get_remote_component_with_mock(
@@ -136,10 +93,7 @@ def test_translate_transformer_lens_path_last_component(adapter: Gemma3Architect
         adapter.translate_transformer_lens_path("embed", last_component_only=True) == "embed_tokens"
     )
     assert adapter.translate_transformer_lens_path("ln_final", last_component_only=True) == "norm"
-    assert (
-        adapter.translate_transformer_lens_path("unembed", last_component_only=True)
-        == "lm_head"
-    )
+    assert adapter.translate_transformer_lens_path("unembed", last_component_only=True) == "lm_head"
 
     # Test block mapping
     assert adapter.translate_transformer_lens_path("blocks", last_component_only=True) == "layers"
